task(SDK - 5523) - Improvements for ZeroBezel
1. Makes MaxLines = 1
2. Makes SmallView for zero bezel Consistent with Basic
3. TextOnly key is redundant now
