+++ clevertap-pushtemplates/src/main/java/com/clevertap/android/pushtemplates/TemplateDataFactory.kt
@@ -39,7 +39,6 @@ import com.clevertap.android.pushtemplates.PTConstants.PT_RENDER_TERMINAL
 import com.clevertap.android.pushtemplates.PTConstants.PT_SCALE_TYPE
 import com.clevertap.android.pushtemplates.PTConstants.PT_SCALE_TYPE_ALT
 import com.clevertap.android.pushtemplates.PTConstants.PT_SCALE_TYPE_COLLAPSED
-import com.clevertap.android.pushtemplates.PTConstants.PT_SMALL_VIEW
 import com.clevertap.android.pushtemplates.PTConstants.PT_STICKY
 import com.clevertap.android.pushtemplates.PTConstants.PT_SUBTITLE
 import com.clevertap.android.pushtemplates.PTConstants.PT_TITLE
@@ -229,7 +228,6 @@ internal object TemplateDataFactory {
             baseContent = createBaseContent(extras, colorMap),
             actions = Utils.getActionKeys(extras),
             mediaData = mediaData,
-            smallView = extras.getString(PT_SMALL_VIEW),
             collapsedMediaData = createCollapsedMediaData(extras, mediaData)
         )
     }
