--- clevertap-pushtemplates/src/main/res/layout/cv_small_text_only.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/content_view_small"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@android:color/transparent"
-    android:paddingLeft="@dimen/padding_horizontal"
-    android:paddingTop="@dimen/padding_micro"
-    android:paddingRight="@dimen/padding_horizontal"
-    android:paddingBottom="@dimen/padding_micro">
-
-    <ImageView
-        android:id="@+id/large_icon"
-        android:layout_width="@dimen/large_icon"
-        android:layout_height="@dimen/large_icon"
-        android:layout_alignParentStart="true"
-        android:layout_marginEnd="@dimen/padding_micro"
-        android:scaleType="centerCrop"/>
-
-
-    <RelativeLayout
-        android:id="@+id/rel_lyt"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toEndOf="@id/large_icon"
-        android:paddingLeft="@dimen/padding_horizontal"
-        android:paddingStart="@dimen/padding_horizontal"
-        android:paddingRight="@dimen/padding_horizontal"
-        android:paddingEnd="@dimen/padding_horizontal">
-
-        <include
-            android:id="@+id/metadata"
-            layout="@layout/metadata"/>
-
-        <TextView
-            android:id="@+id/title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/metadata"
-            android:ellipsize="end"
-            android:maxLines="2"
-            android:layout_marginTop="@dimen/padding_micro"
-            tools:text="title"
-            android:textAppearance="@style/PushTitle"/>
-    </RelativeLayout>
-
-</RelativeLayout>
