+++ clevertap-pushtemplates/src/main/res/layout/cv_small_zero_bezel.xml
@@ -22,8 +22,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentStart="true"
-        android:layout_toStartOf="@+id/large_icon"
-        android:layout_alignBottom="@+id/big_media_configurable"
+        android:layout_centerVertical="true"
         android:paddingLeft="@dimen/padding_horizontal"
         android:paddingStart="@dimen/padding_horizontal"
         android:paddingRight="@dimen/padding_horizontal"
@@ -41,6 +40,7 @@
             android:layout_height="wrap_content"
             android:layout_below="@+id/metadata"
             android:layout_alignStart="@+id/metadata"
+            android:layout_toStartOf="@+id/large_icon"
             android:layout_marginTop="@dimen/padding_micro"
             android:ellipsize="end"
             android:maxLines="1"
@@ -53,20 +53,21 @@
             android:layout_height="wrap_content"
             android:layout_below="@+id/title"
             android:layout_alignStart="@id/title"
+            android:layout_toStartOf="@+id/large_icon"
             android:ellipsize="end"
-            android:maxLines="2"
+            android:maxLines="1"
             tools:text="message message message"
             android:textAppearance="@style/PushMessage"/>
 
-    </RelativeLayout>
+        <ImageView
+            android:id="@+id/large_icon"
+            android:layout_width="@dimen/large_icon"
+            android:layout_height="@dimen/large_icon"
+            android:layout_marginStart="@dimen/metadata_title_margin_horizontal"
+            android:layout_alignParentEnd="true"
+            android:layout_centerInParent="true"
+            android:importantForAccessibility="no"
+            android:scaleType="centerCrop" />
 
-    <ImageView
-        android:id="@+id/large_icon"
-        android:layout_alignBottom="@+id/big_media_configurable"
-        android:layout_width="@dimen/large_icon"
-        android:layout_height="@dimen/large_icon"
-        android:layout_marginStart="@dimen/metadata_title_margin_horizontal"
-        android:layout_marginBottom="@dimen/padding_micro"
-        android:layout_alignParentEnd="true"
-        android:scaleType="centerCrop" />
+    </RelativeLayout>
 </RelativeLayout>
